moved traits to mod.rs
